refactor(ui): split mounted diff row views
